refactor: delegate heavy processing to services, keeping subscribers lightweight
